requirements.txt
absl_py==1.4.0
absl_py==2.0.0
crc32c==2.8
h5py==3.9.0
imageio==2.21.3
loguru==0.5.3
matplotlib==3.10.7
numba==0.58.0
numpy==1.25.0
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==12.0.0
protobuf==3.20.1
protobuf==3.20.3
scikit_learn==1.3.0
scipy==1.16.2
six==1.16.0
skimage==0.0
torchvision==0.16.0
tqdm==4.65.0
Wand==0.6.11
torch==2.1.0